Fix FlexDevice.h internal function signature and update .gitignore
Fixed compilation issue:
- _flex_build_cmd_send_flex: uint32_t capcode → uint64_t capcode
- This internal function was missed in the initial uint64_t migration

Updated .gitignore:
- Added cli/flex-cli (compiled binary)
- Added cli/libraries/gcc/example/example_gcc (compiled binary)
- Added cli/libflex_binary (symlink to libraries/gcc)
- Added __pycache__/ and *.pyc (Python cache files)

Compilation verified:
- flex-cli compiles and runs: v2.5.5
- example_gcc compiles and runs successfully
- Python example runs successfully with PYTHONPATH

All clients now ready for testing with full capcode range support.

+++ cli/libraries/gcc/FlexDevice.h
@@ -411,7 +411,7 @@ static inline void _flex_populate_ts(uint8_t *raw) {
 
 static inline void _flex_build_cmd_send_flex(uint8_t raw[FLEX_PACKET_SIZE],
                                               uint8_t seq, const uint8_t uuid[16],
-                                              uint32_t capcode, float frequency,
+                                              uint64_t capcode, float frequency,
                                               int8_t power, uint8_t mail_drop,
                                               const char *message, uint8_t msg_len) {
     memset(raw, 0, FLEX_PACKET_SIZE);
